[REPO] Add wrap IR from hnimast back ... again
Originally I moved implementation to hnimast because I thought it would
make sense to put `cgen`, `wrapheader` and other macros into library that
does not depend on htsparse etc. - `hcparse` was not supposed to be used as
a direct dependency. Generated wrappers don't depend on anything except
some things from `hmisc/wrappers/wraphelp` (and this could be moved in a
separate module sometimes later if needed). But `interop_ir` got /a little/
more complex than I originally anticipated, and managing it separately is
very annoying.

Ideally problem like this is should be solved by tooling, and not by me
shuffling code around and fragmenting packages. This can be solved by
support for 'package features'.