[WIP] Refactor old libclang-based implementation ot use `CxxT` IR
Remove old implementation based on the `GenT` and use new `CxxT` IR. This
is needed for multiple reasons - first of all, old implementation was in
need of total refactoring, because it had a lot of overlapping
functionality that was reimplemented, then forgotten and reimplemented
again, when I had another round of refactoring and fixes. Second issue -
`GenT` was an unnecessary overcommplication that is better replaced with
`CxxT` that was designed from the start, rather then being gradually
expanded based on new requirements. They do share large portion of the
structure, but type and name handling is different.