[WIP] Porting C++ type graph grouping.
Currently compilation fails on the line 545 of 1122, which is approximately
~49%. Good enough for not, will continue fixing it later.

So far the idea of moving to `CxxT` for all operations seems fine. There
are of course some things that need to be implemented from scratch, like
handling of the default template parameters, but that is mainly related to
horrible state of the existing implementation.

Testing new solution will be much easier - first of all I can use
tree-sitter based solution to generate `CxxT` entries from simple string,
and generate tests on lots of small examples. Testing separate parts of the
pipeline is much easier as well - earlier graph construction relied too
much on the cursor location information, that could not be easily generated
from scratch, which meant I had to test things on the real code. And so on.